first official

+++ Screens/Main/Profile/EditProperty/editProperty.js
@@ -12,7 +12,7 @@ import {
   Alert
 } from 'react-native';
 
-import SecureStorage, { ACCESS_CONTROL, ACCESSIBLE, AUTHENTICATION_TYPE } from 'react-native-secure-storage'
+import EncryptedStorage from 'react-native-encrypted-storage';
 
 import { HEIGHT, WIDTH, PRIMARYCOLOR, DARKGREY, LIGHTGREY} from '../../../../sharedUtils'
 
@@ -30,6 +30,7 @@ import { HeaderContainer, BackButtonContainer, NameContainer, Header, ResetButto
 import { ScrollView } from 'react-native-gesture-handler';
 
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { acc } from 'react-native-reanimated';
 
 export default function EditPropertyScreen({navigation, route}){
 
@@ -56,128 +57,143 @@ export default function EditPropertyScreen({navigation, route}){
 
 
     async function getTokens(){
-        console.log("refresh")
-        const accessToken = await SecureStorage.getItem("accessToken");
-        const UID = await SecureStorage.getItem("userId");
 
-        fetch('https://crib-llc.herokuapp.com/properties/' + route.params.propId, {
-        method: 'POST',
-        headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + accessToken,
+        try{
+            const accessToken = await EncryptedStorage.getItem("accessToken");
+
+            if(accessToken != undefined){
+                fetch('https://crib-llc.herokuapp.com/properties/' + route.params.propId, {
+                method: 'POST',
+                headers: {
+                Accept: 'application/json',
+                'Content-Type': 'application/json',
+                'Authorization': 'Bearer ' + accessToken,
+                }
+                }) 
+                .then(res => res.json()).then(async propData =>{
+                    setPropAPIData(propData) 
+
+                    //Set the default varaibles 
+                    if(route.params.propertyData.type == undefined || propType != propData.propertyInfo.type){
+                        // console.log("UPDATE --- API --- propType")
+                        setPropType(propData.propertyInfo.type)
+                    }
+                    else{
+                        // console.log("UPDATE --- PARAMS --- propType")
+                    }
+                    if(route.params.propertyData.price == undefined || propPrice != propData.propertyInfo.price){
+                        // console.log("UPDATE --- API --- propPrice")
+                        setPropPrice(propData.propertyInfo.price)
+                    }
+                    else{
+                        // console.log("UPDATE --- PARAMS --- propPrice")
+                    }
+                    if(route.params.propertyData.propDateFrom == undefined || propDateFrom != propData.propertyInfo.availableFrom){
+                        // console.log("UPDATE --- API --- propavailFrom")
+                        setPropDateFrom(propData.propertyInfo.availableFrom)
+                    }
+                    else{
+                        // console.log("UPDATE --- PARAMS --- propavailableFrom")
+                    }
+                    if(route.params.propertyData.propDateTo == undefined != propData.propertyInfo.availableTo){
+                        // console.log("UPDATE --- API --- propavailTo")
+                        setPropDateTo(propData.propertyInfo.availableTo)
+                    }
+                    else{
+                        // console.log("UPDATE --- PARAMS --- propavailableTo")
+                    }
+                    if(route.params.propertyData.description == undefined || propDescription != propData.propertyInfo.description){
+                        // console.log("UPDATE --- API --- propDescription")
+                        setPropDescription(propData.propertyInfo.description)
+                    }
+                    else{
+                        // console.log("UPDATE --- PARAMS --- propDescription")
+                    }
+                    if(route.params.propertyData._id == undefined ||  propID != propData.propertyInfo._id){
+                        // console.log("UPDATE --- API --- propID")
+                        setPropID(propData.propertyInfo._id)
+                    }
+                    else{
+                        // console.log("UPDATE --- PARAMS --- propID")
+                    }
+                    if(route.params.propertyData.amenities == undefined || new Object(propAmen).toLocaleString() != new Object(propData.propertyInfo.amenities).toLocaleString()){
+                        // console.log("UPDATE --- API --- propAmen")
+                        setPropAmen(propData.propertyInfo.amenities)
+                    }
+                    else{
+                        // console.log("UPDATE --- PARAMS --- propAmen")
+                    }
+                    if(route.params.propertyData.imgList == undefined || new Object(propImg).toLocaleString() != new Object(propData.propertyInfo.imgList).toLocaleString()){
+                        // console.log("UPDATE --- API --- propImg")
+                        setPropImg(propData.propertyInfo.imgList)
+                    }
+                    else{
+                        // console.log("UPDATE --- PARAMS --- propImg")
+                    }
+                    //console.log(propData.propertyInfo.imgList)
+
+                })
+                .catch(e=>{
+                    alert(e)
+                })
+            }
         }
-        }) 
-        .then(res => res.json()).then(async propData =>{
-            setPropAPIData(propData) 
+        catch{
 
-            //Set the default varaibles 
-            if(route.params.propertyData.type == undefined || propType != propData.propertyInfo.type){
-                console.log("UPDATE --- API --- propType")
-                setPropType(propData.propertyInfo.type)
-            }
-            else{
-                console.log("UPDATE --- PARAMS --- propType")
-            }
-            if(route.params.propertyData.price == undefined || propPrice != propData.propertyInfo.price){
-                console.log("UPDATE --- API --- propPrice")
-                setPropPrice(propData.propertyInfo.price)
-            }
-            else{
-                console.log("UPDATE --- PARAMS --- propPrice")
-            }
-            if(route.params.propertyData.propDateFrom == undefined || propDateFrom != propData.propertyInfo.availableFrom){
-                console.log("UPDATE --- API --- propavailFrom")
-                setPropDateFrom(propData.propertyInfo.availableFrom)
-            }
-            else{
-                console.log("UPDATE --- PARAMS --- propavailableFrom")
-            }
-            if(route.params.propertyData.propDateTo == undefined != propData.propertyInfo.availableTo){
-                console.log("UPDATE --- API --- propavailTo")
-                setPropDateTo(propData.propertyInfo.availableTo)
-            }
-            else{
-                console.log("UPDATE --- PARAMS --- propavailableTo")
-            }
-            if(route.params.propertyData.description == undefined || propDescription != propData.propertyInfo.description){
-                console.log("UPDATE --- API --- propDescription")
-                setPropDescription(propData.propertyInfo.description)
-            }
-            else{
-                console.log("UPDATE --- PARAMS --- propDescription")
-            }
-            if(route.params.propertyData._id == undefined ||  propID != propData.propertyInfo._id){
-                console.log("UPDATE --- API --- propID")
-                setPropID(propData.propertyInfo._id)
-            }
-            else{
-                console.log("UPDATE --- PARAMS --- propID")
-            }
-            if(route.params.propertyData.amenities == undefined || new Object(propAmen).toLocaleString() != new Object(propData.propertyInfo.amenities).toLocaleString()){
-                console.log("UPDATE --- API --- propAmen")
-                setPropAmen(propData.propertyInfo.amenities)
-            }
-            else{
-                console.log("UPDATE --- PARAMS --- propAmen")
-            }
-            if(route.params.propertyData.imgList == undefined || new Object(propImg).toLocaleString() != new Object(propData.propertyInfo.imgList).toLocaleString()){
-                console.log("UPDATE --- API --- propImg")
-                setPropImg(propData.propertyInfo.imgList)
-            }
-            else{
-                console.log("UPDATE --- PARAMS --- propImg")
-            }
-            //console.log(propData.propertyInfo.imgList)
-
-        })
-        .catch(e=>{
-            alert(e)
-        })
+        }
     }
 
     async function SelectPropPic(index){
-       
-        const accessToken = await SecureStorage.getItem("accessToken");
-        ImagePicker.openPicker({
-            width: 812,
-            height: 812,
-            cropping:true,
-            compressImageQuality: 0.8
-          }).then(image => {
+       try{
+        const accessToken = await EncryptedStorage.getItem("accessToken");
+        if(accessToken != undefined){
+            const accessToken = await EncryptedStorage.getItem("accessToken");
+            ImagePicker.openPicker({
+                width: 812,
+                height: 812,
+                cropping:true,
+                compressImageQuality: 0.8
+            }).then(image => {
 
-            const formData = new FormData();
-            var array = image.path.split(".");      
-            formData.append("propertyImage" , {
-                uri: image.path,
-                type: 'image/' + array[1],
-                name: 'someName',
-            }); 
-            formData.append("changeIdx", index);
-           
-           
-            fetch('https://crib-llc.herokuapp.com/properties/propertyImages/' + propID, {
-                method: 'PUT',
-                headers: {
-                    Accept: 'application/json',
-                    'Content-Type': 'application/json',
-                    'Authorization': 'Bearer ' + accessToken,
-                },
-                body: formData
+                const formData = new FormData();
+                var array = image.path.split(".");      
+                formData.append("propertyImage" , {
+                    uri: image.path,
+                    type: 'image/' + array[1],
+                    name: 'someName',
+                }); 
+                formData.append("changeIdx", index);
+            
+            
+                fetch('https://crib-llc.herokuapp.com/properties/propertyImages/' + propID, {
+                    method: 'PUT',
+                    headers: {
+                        Accept: 'application/json',
+                        'Content-Type': 'application/json',
+                        'Authorization': 'Bearer ' + accessToken,
+                    },
+                    body: formData
+                })
+                .then(res => res.json()).then(async data=>{
+                    setPropImg([...propImg.slice(0, index),data.propertyImage,...propImg.slice(index + 1),])
+                    try{
+                        await AsyncStorage.removeItem('postedProperty')
+                    }
+                    catch{
+                        alert("An error has occured. Please try again later!")
+                    }
+                    
+                })
+                .catch((error) => {
+                    console.log(error)
+                });
+            }).catch(e=>{
             })
-            .then(res => res.json()).then(async data=>{
-                setPropImg([...propImg.slice(0, index),data.propertyImage,...propImg.slice(index + 1),])
-                await AsyncStorage.removeItem('postedProperty')
-                console.log(data)
-            })
-            .catch((error) => {
-                console.log(error)
-            });
-        }).catch(e=>{
-            console.log("Canceled")
-        })
-        
-        
+        }
+       }
+       catch{
+        console.log("ERROR --- SEELCTPIC")
+       }
     }   
 
 
@@ -195,27 +211,36 @@ export default function EditPropertyScreen({navigation, route}){
           );
     }
     async function deletePropertyRequest(){
-        const accessToken = await SecureStorage.getItem("accessToken");
-        fetch('https://crib-llc.herokuapp.com/properties/' + propID, {
-            method: 'DELETE',
-            headers: {
-                Accept: 'application/json',
-                'Content-Type': 'application/json',
-                'Authorization': 'Bearer ' + accessToken,
+        try{
+            const accessToken = await EncryptedStorage.getItem("accessToken");
+            if(accessToken != undefined && propID != null && propID != undefined){
+                fetch('https://crib-llc.herokuapp.com/properties/' + propID, {
+                    method: 'DELETE',
+                    headers: {
+                        Accept: 'application/json',
+                        'Content-Type': 'application/json',
+                        'Authorization': 'Bearer ' + accessToken,
+                    }
+                }).then(res => {
+                    if(res.status == 200){
+
+                            navigation.goBack()
+                        }
+                        else{
+                            alert('Unable to delete this property. Please try again later.')
+                        }
+                    
+                })
+                .catch((error) => {
+
+                });
             }
-        }).then(res => {
-            if(res.status == 200){
-
-                    navigation.goBack()
-                }
-                else{
-                    alert('Unable to delete this property. Please try again later.')
-                }
-            
-        })
-        .catch((error) => {
-
-        });
+        }
+        catch{
+            console.log("ERROR --- DELETEPROPERTYREQUEST")
+        }
+        
+        
     }   
     
     return(
@@ -223,7 +248,7 @@ export default function EditPropertyScreen({navigation, route}){
         <SafeAreaView style={{flex:1, backgroundColor:'white'}} >
             <HeaderContainer>
                 <BackButtonContainer>
-                    <Pressable style={{height:'50%', width:'50%', alignItems:'center'}} onPress={()=> navigation.goBack()}>
+                    <Pressable hitSlop={WIDTH*0.025} onPress={()=> navigation.goBack()}>
                         <Ionicons name='arrow-back-outline' size={25} style={{paddingHorizontal:WIDTH*0.02}}/>
                     </Pressable>
                 </BackButtonContainer>
@@ -308,10 +333,10 @@ export default function EditPropertyScreen({navigation, route}){
             <CategoryName>Availability</CategoryName>
             <RowContainer onPress={()=> navigation.navigate("EditPropertyAvail",{from: propDateFrom, to: propDateTo, uid: propID, propertyData: route.params.propertyData})}>
                 <DatePriceText>
-                    {new Date(propDateFrom).getUTCMonth()% 12 + 1}- 
+                    {new Date(propDateFrom).toDateString().split(" ")[1]} {new Date(propDateFrom).toDateString().split(" ")[2]}{" "}
                     {new Date(propDateFrom).getFullYear()}
                     {" "} to {" "}
-                    {new Date(propDateTo).getUTCMonth() % 12 + 1}- 
+                    {new Date(propDateTo).toDateString().split(" ")[1]} {new Date(propDateTo).toDateString().split(" ")[2]}{" "}
                     {new Date(propDateTo).getFullYear()}
                 </DatePriceText>
                 <Ionicons name='chevron-forward-outline' size={25}  style={{paddingLeft: WIDTH*0.05}}/>
